use middlewares/auth/{test,supabase,firebase}
