+++ readme.md
@@ -36,7 +36,7 @@ The project requires access to Supabase and Firebase databases.
 
 To get the needed keys for local development, run `vercel env pull .env.local`.
 
-Preview/Staging and Production keys are only defined in the Vercel project. 
+Preview/Staging and Production keys are only defined in the Vercel project.
 
 ## Deployment to production
 
@@ -51,3 +51,9 @@ You can start editing the page by modifying `pages/index.js`. The page auto-upda
 [API routes](https://nextjs.org/docs/api-routes/introduction) can be accessed on [http://localhost:3000/api/hello](http://localhost:3000/api/hello). This endpoint can be edited in `pages/api/hello.js`.
 
 The `pages/api` directory is mapped to `/api/*`. Files in this directory are treated as [API routes](https://nextjs.org/docs/api-routes/introduction) instead of React pages.
+
+# TESTS
+To test the endpoints, use this:
+```
+	curl -X POST http://localhost:3000/import/firebase -d '{"tokenFirebase":"value1", "tokenSupabase":"value2"}' -H "Content-Type: application/json" -i
+```
