+++ utils/supabase-admin.js
@@ -5,19 +5,8 @@ const {SUPABASE_URL, SUPABASE_ANON_KEY} = config
 
 const supabaseClient = createClient(SUPABASE_URL, SUPABASE_ANON_KEY)
 
-// require supabase authentication
-export const requireSupabaseSession = (fn) => async (req, res) => {
-	const user = await verifySupabaseToken(req.body.tokenSupabase)
-	if (!user) return res.status(401).json({
-		message: 'Not signed in to supabase; ?tokenSupabase='
-	})
-	req.user = user
-	return fn(req, res)
-}
-
 export const verifySupabaseToken = async (access_token) => {
-	const {user} = await supabaseClient.auth.api.getUser(access_token)
-	return user
+	return supabaseClient.auth.api.getUser(access_token)
 }
 
 export default supabaseClient
